requirements.txt
chromadb==0.4.15
pydantic==2.5.2
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-groq==0.0.1
langchain-google-genai==0.0.5
langchain-text-splitters==0.0.1
pypdf==4.0.1
streamlit==1.33.0
pandas>=2.0.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
